소요시간 분/초 정수로만 입력 가능하도록 수정

+++ src/frontend/src/pages/content-wage-list/tabs/content-wage-table-tab/components/user-content-duration-edit-dialog.tsx
@@ -13,8 +13,8 @@ import {
 
 const schema = z.object({
   id: z.number(),
-  minutes: z.number().min(0),
-  seconds: z.number().min(0).max(59),
+  minutes: z.number().int().min(0),
+  seconds: z.number().int().min(0).max(59),
 });
 
 type UserContentDurationEditDialogProps = {
